+++ tests/test_utils.py
@@ -39,9 +39,9 @@ from dwave.cloud.utils.qubo import (
     active_qubits, generate_random_ising_problem)
 from dwave.cloud.utils.time import utcnow
 from dwave.cloud.utils.cli import default_text_input
+from dwave.cloud.utils.decorators import aliasdict, cached, deprecated, retried
 from dwave.cloud.utils import (
     NumpyEncoder, coerce_numpy_to_python,
-    cached, retried, deprecated, aliasdict,
     hasinstance, exception_chain,
     is_caused_by, get_distribution, PackageNotFoundError, VersionNotFoundError)
